avr: Separate juksautus specific functions out of main #10
